+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 #### = 1.0.9 (2025-09-20) =
 
 * added results history view and editor
+* show time in list if set
+* added results and recuperation symbol in list
 
 #### = 1.0.8 (2025-09-07) =